Implement multicolumn primary keys
This is a part of #23.

This is first step in introduction multicolumn primary keys. This change
introduces a primary key as a collection of several columns with
limitation of support for i64 type only. It changes HTTP API according
to this change also. Internally it use only one column in the primary
key collection - the one which is retrieved from IndexMetadata.