benchmark: add build-index command
This patch implements creating vector search index in the ScyllaDB and measures
the operation time.  The operation will start once the index in ScyllaDB is
created, it will finishThe operation will start once the index in ScyllaDB is
created, it will finish when all all Vector Store services returns index size
equal to the dataset size.