vector-store: add tests for ann filtering
The commit adds tests for ann filtering for int primary key and eq, eq tuple
and IN restrictions.

The commit fixes db_basic mock to be able to process messages while sending
values to the embedding stream.

The commit changes default space type for integration tests to Euclidean as it
seems to be easier for a human to prepare test data.